refactor(core): move path constants to lionagi._paths, fix ops-plane→cli imports
LIONAGI_HOME already lived in _paths.py; this slice adds RUNS_ROOT and
find_lionagi_dirs/_find_lionagi_dirs so the ops plane (studio/services,
state/provenance) can import path helpers from the leaf module instead of
reaching up into cli/_runs or cli/_agents.

cli/_runs.py and cli/_agents.py re-export the moved symbols under their
original names so all cli-internal callers remain unchanged. Tests that
monkeypatched cli._agents._find_lionagi_dirs are updated to patch
lionagi._paths.find_lionagi_dirs, which is the actual source.
