+++ lionagi/state/provenance.py
@@ -29,9 +29,10 @@ def agent_definition_hash(agent_name: str | None) -> str | None:
     """
     if not agent_name:
         return None
-    from lionagi.cli._agents import _find_lionagi_dirs, _resolve_profile_path
+    from lionagi._paths import find_lionagi_dirs
+    from lionagi.cli._agents import _resolve_profile_path
 
-    for d in _find_lionagi_dirs():
+    for d in find_lionagi_dirs():
         path = _resolve_profile_path(d / "agents", agent_name)
         if path is not None:
             return _hash_file(path)
