+++ lionagi/studio/services/agents.py
@@ -5,7 +5,7 @@ from typing import Any
 
 import yaml
 
-from lionagi.cli._runs import LIONAGI_HOME
+from lionagi._paths import LIONAGI_HOME
 from lionagi.libs.frontmatter import parse_frontmatter as _parse_frontmatter
 
 from ._path_safety import public_path, safe_path_join
