+++ lionagi/studio/services/skills.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from lionagi.cli._runs import LIONAGI_HOME
+from lionagi._paths import LIONAGI_HOME
 from lionagi.libs.frontmatter import parse_frontmatter as _parse_frontmatter
 
 from ._path_safety import public_path, safe_path_join
